# projects/workplane/contracts/openapi.yaml
openapi: 3.1.0
...
info:
  title: Workplane Planning Contract API
  version: 0.5.0-m2c
  description: Executable M1/M2 durable realtime behavior plus the M2C planning-contract spine.
servers:
  - url: http://localhost:8080
paths:
  /api/v1/session/login:
    post:
      operationId: login
      security: []
      x-permission-actions: [session.login]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LoginRequest'}
      responses:
        '200':
          description: Authenticated human session
          headers:
            Set-Cookie:
              description: >-
                Issues the HttpOnly workplane_session credential. Secure is
                required outside explicit localhost development; Path=/ and
                SameSite=Lax are always required.
              required: true
              schema:
                type: string
                pattern: '^workplane_session=[^;]+; Path=/; HttpOnly; SameSite=Lax(?:; Secure)?$'
                examples:
                  - 'workplane_session=opaque; Path=/; HttpOnly; SameSite=Lax; Secure'
            X-Request-ID: {$ref: '#/components/headers/RequestId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Session'}
        default: {$ref: '#/components/responses/Problem'}
  /api/v1/orgs/{org_id}/projects:
    parameters:
      - {$ref: '#/components/parameters/OrgId'}
    post:
      operationId: createProject
      description: >-
        Creates an organization-level exploration project. An agent bearer may
        call this operation only when its project allowlist is empty; any
        nonempty project restriction denies creation because no existing
        project can be the operation target.
      security:
        - HumanSession: []
          CsrfToken: []
        - AgentBearer: []
      x-permission-actions: [project.create]
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateExplorationProject'}
      responses:
        '201':
          description: Exploration project created
          headers:
            X-Request-ID: {$ref: '#/components/headers/RequestId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Project'}
        default: {$ref: '#/components/responses/Problem'}
  /api/v1/projects/{project_id}:
    parameters:
      - {$ref: '#/components/parameters/ProjectId'}
    get:
      operationId: getProject
      security:
        - HumanSession: []
        - AgentBearer: []
      x-permission-actions: [project.read]
      responses:
        '200':
          description: Project projection
          headers:
            X-Request-ID: {$ref: '#/components/headers/RequestId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Project'}
        default: {$ref: '#/components/responses/Problem'}
  /api/v1/projects/{project_id}/activate:
    parameters: [{$ref: '#/components/parameters/ProjectId'}]
    post:
      operationId: activateProject
      security: &mutationSecurity
        - HumanSession: []
          CsrfToken: []
        - AgentBearer: []
      x-permission-actions: [project.activate]
      parameters: &versionedMutationParameters
        - {$ref: '#/components/parameters/IdempotencyKey'}
        - {$ref: '#/components/parameters/ExpectedVersion'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LifecycleReasonRequest'}
      responses:
        '200': &versionedProjectResponse
          description: Committed project projection
          headers:
            ETag: {$ref: '#/components/headers/VersionETag'}
            X-Request-ID: {$ref: '#/components/headers/RequestId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Project'}
        default: {$ref: '#/components/responses/Problem'}
  /api/v1/projects/{project_id}/hold:
    parameters: [{$ref: '#/components/parameters/ProjectId'}]
    post:
      operationId: holdProject
      security: *mutationSecurity
      x-permission-actions: [project.hold, decision.record]
      parameters: *versionedMutationParameters
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LifecycleReasonRequest'}
      responses:
        '200': *versionedProjectResponse
        default: {$ref: '#/components/responses/Problem'}
  /api/v1/projects/{project_id}/resume:
    parameters: [{$ref: '#/components/parameters/ProjectId'}]
    post:
      operationId: resumeProject
      security: *mutationSecurity
      x-permission-actions: [project.resume, decision.record]
      parameters: *versionedMutationParameters
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LifecycleReasonRequest'}
      responses:
        '200': *versionedProjectResponse
        default: {$ref: '#/components/responses/Problem'}
  /api/v1/projects/{project_id}/promote:
    parameters: [{$ref: '#/components/parameters/ProjectId'}]
    post:
      operationId: promoteProject
      security: *mutationSecurity
      x-permission-actions: [project.promote, decision.record]
      parameters: *versionedMutationParameters
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PromoteProjectRequest'}
      responses:
        '200':
          description: Atomic promotion result
          headers:
            ETag: {$ref: '#/components/headers/VersionETag'}
            X-Request-ID: {$ref: '#/components/headers/RequestId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PromotionResult'}
        default: {$ref: '#/components/responses/Problem'}
...
  /api/v1/projects/{project_id}/forecasts:
    parameters: [{$ref: '#/components/parameters/ProjectId'}]
    get:
      operationId: listProjectForecasts
      security: *readSecurity
      x-permission-actions: [project.read]
      responses:
        '200': &forecastHistoryResponse
          description: Immutable forecast history with current and stale projection flags
          headers:
            X-Request-ID: {$ref: '#/components/headers/RequestId'}
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/Forecast'}}
        default: {$ref: '#/components/responses/Problem'}
    post:
      operationId: reforecastProject
      security: *mutationSecurity
      x-permission-actions: [project.reforecast]
      parameters: *versionedMutationParameters
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ForecastInput'}
      responses:
        '201': &versionedForecastResponse
          description: Immutable current forecast
          headers:
            ETag: {$ref: '#/components/headers/VersionETag'}
            X-Request-ID: {$ref: '#/components/headers/RequestId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Forecast'}
        default: {$ref: '#/components/responses/Problem'}
...
  /api/v1/projects/{project_id}/target:
    parameters: [{$ref: '#/components/parameters/ProjectId'}]
    post:
      operationId: setProjectTarget
      security: *mutationSecurity
      x-permission-actions: [project.target.write]
      parameters: *versionedMutationParameters
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/TargetInput'}
      responses:
        '201': &versionedTargetResponse
          description: Immutable current intent target
          headers:
            ETag: {$ref: '#/components/headers/VersionETag'}
            X-Request-ID: {$ref: '#/components/headers/RequestId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Target'}
        default: {$ref: '#/components/responses/Problem'}
  /api/v1/projects/{project_id}/deadline:
    parameters: [{$ref: '#/components/parameters/ProjectId'}]
    post:
      operationId: setProjectDeadline
      security: *mutationSecurity
      x-permission-actions: [project.deadline.write]
      parameters: *versionedMutationParameters
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/DeadlineInput'}
      responses:
        '201':
          description: Immutable current typed deadline
          headers:
            ETag: {$ref: '#/components/headers/VersionETag'}
            X-Request-ID: {$ref: '#/components/headers/RequestId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Deadline'}
        default: {$ref: '#/components/responses/Problem'}
  /api/v1/projects/{project_id}/decisions:
    parameters:
      - {$ref: '#/components/parameters/ProjectId'}
    post:
      operationId: recordDecision
      security:
        - HumanSession: []
          CsrfToken: []
        - AgentBearer: []
      x-permission-actions: [decision.record]
      parameters:
        - {$ref: '#/components/parameters/IdempotencyKey'}
        - {$ref: '#/components/parameters/ExpectedVersion'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RecordDecision'}
      responses:
        '201':
          description: Decision recorded
          headers:
            ETag:
              description: Quoted aggregate version returned by the committed mutation.
              required: true
              schema: {$ref: '#/components/schemas/VersionETag'}
            X-Request-ID: {$ref: '#/components/headers/RequestId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Decision'}
        default: {$ref: '#/components/responses/Problem'}
  /api/v1/projects/{project_id}/activity:
    parameters:
      - {$ref: '#/components/parameters/ProjectId'}
    get:
      operationId: listProjectActivity
      security:
        - HumanSession: []
        - AgentBearer: []
      x-permission-actions: [project.read]
      responses:
        '200':
          description: Immutable activity entries
          headers:
            X-Request-ID: {$ref: '#/components/headers/RequestId'}
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Activity'}
        default: {$ref: '#/components/responses/Problem'}
...
components:
  headers:
    RequestId:
      description: Server-generated identifier for the request and committed command.
      required: true
      schema: {type: string, minLength: 1, maxLength: 128}
    VersionETag:
      description: Quoted aggregate version returned by the committed mutation.
      required: true
      schema: {$ref: '#/components/schemas/VersionETag'}
  securitySchemes:
    HumanSession:
      type: apiKey
      in: cookie
      name: workplane_session
      description: HttpOnly, Secure, SameSite session cookie for human browser requests.
    CsrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Required with the human session cookie for every protected mutation.
    AgentBearer:
      type: http
      scheme: bearer
      bearerFormat: WorkplaneAgentToken
      description: >-
        Scoped agent token carrying the agent identity and delegated human
        principal. An empty project allowlist is organization-scoped. A
        nonempty allowlist is project-scoped: only listed concrete projects
        are eligible, and organization-level project creation is denied.
  parameters:
    OrgId:
      name: org_id
      in: path
      required: true
      schema: {type: string, format: uuid}
    ProjectId:
      name: project_id
      in: path
      required: true
      schema: {type: string, format: uuid}
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 16, maxLength: 128}
    ExpectedVersion:
      name: If-Match
      in: header
      required: true
      description: Quoted current aggregate version; stale or missing values deny before mutation.
      schema: {$ref: '#/components/schemas/VersionETag'}
...
  responses:
    Problem:
      description: Typed problem response
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
...
  schemas:
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email: {type: string, format: email, minLength: 1, maxLength: 320}
        password: {type: string, minLength: 1, maxLength: 1024}
    Session:
      type: object
      additionalProperties: false
      required: [actor_id, actor_kind, expires_at, csrf_token]
      properties:
        actor_id: {type: string, format: uuid}
        actor_kind: {type: string, enum: [human]}
        expires_at: {type: string, format: date-time}
        csrf_token:
          type: string
          minLength: 32
          maxLength: 256
          description: Submit as X-CSRF-Token with the issued session cookie on protected mutations.
    CreateExplorationProject:
      type: object
      additionalProperties: false
      required: [title, outcome, hypothesis, falsifier, decision_criteria, experiment_bound]
      properties:
        title: {type: string, minLength: 1, maxLength: 200, pattern: '\S'}
        outcome: {type: string, minLength: 1, maxLength: 2000, pattern: '\S'}
        hypothesis: {type: string, minLength: 1, maxLength: 2000, pattern: '\S'}
        falsifier: {type: string, minLength: 1, maxLength: 2000, pattern: '\S'}
        decision_criteria:
          type: array
          minItems: 1
          maxItems: 32
          items: {type: string, minLength: 1, maxLength: 500, pattern: '\S'}
        experiment_bound: {type: string, minLength: 1, maxLength: 1000, pattern: '\S'}
    Project:
      type: object
      additionalProperties: false
      required: [id, organization_id, title, outcome, mode, state, version, hypothesis, falsifier, decision_criteria, experiment_bound]
      properties:
        id: {type: string, format: uuid}
        organization_id: {type: string, format: uuid}
        title: {type: string, minLength: 1, maxLength: 200, pattern: '\S'}
        outcome: {type: string, minLength: 1, maxLength: 2000, pattern: '\S'}
        mode: {type: string, enum: [exploration, exploitation]}
        state: {type: string, enum: [proposed, active, held, completed, stopped, cancelled]}
        version: {type: integer, minimum: 1}
        hypothesis: {type: string, minLength: 1, maxLength: 2000, pattern: '\S'}
        falsifier: {type: string, minLength: 1, maxLength: 2000, pattern: '\S'}
        decision_criteria:
          type: array
          minItems: 1
          maxItems: 32
          items: {type: string, minLength: 1, maxLength: 500, pattern: '\S'}
        experiment_bound: {type: string, minLength: 1, maxLength: 1000, pattern: '\S'}
    RecordDecision:
      type: object
      additionalProperties: false
      required: [kind, question, choice, alternatives, rationale, evidence, consequences]
      properties:
        kind: {type: string, enum: [continue]}
        question: {type: string, minLength: 1, maxLength: 2000, pattern: '\S'}
        choice: {type: string, minLength: 1, maxLength: 2000, pattern: '\S'}
        alternatives:
          type: array
          maxItems: 32
          items: {type: string, maxLength: 1000}
        rationale: {type: string, minLength: 1, maxLength: 4000, pattern: '\S'}
        evidence:
          type: array
          maxItems: 64
          items: {type: string, maxLength: 2000}
        consequences:
          type: array
          maxItems: 32
          items: {type: string, maxLength: 2000}
    LifecycleReasonRequest:
      type: object
      additionalProperties: false
      required: [reason]
      properties:
        reason: {type: string, minLength: 1, maxLength: 4000, pattern: '\S'}
...
    PromoteProjectRequest:
      type: object
      additionalProperties: false
      required: [decision, residual_uncertainty, priority_rationale, deliverables]
      properties:
        decision: {$ref: '#/components/schemas/PromotionDecision'}
        residual_uncertainty: {type: string, minLength: 1, maxLength: 4000, pattern: '\S'}
        priority_rationale: {type: string, minLength: 1, maxLength: 4000, pattern: '\S'}
        deliverables:
          type: array
          minItems: 1
          maxItems: 32
          items: {$ref: '#/components/schemas/DeliverableInput'}
    ForecastInput:
      type: object
      additionalProperties: false
      required: [p50_at, p90_at, review_after, basis, assumptions, reason_codes, impact]
      properties:
        p50_at: {type: string, format: date-time}
        p90_at: {type: string, format: date-time}
        review_after: {type: string, format: date-time}
        basis: {type: string, minLength: 1, maxLength: 4000, pattern: '\S'}
        assumptions:
          type: array
          minItems: 1
          maxItems: 32
          items: {type: string, minLength: 1, maxLength: 1000, pattern: '\S'}
        reason_codes:
          type: array
          minItems: 1
          maxItems: 9
          uniqueItems: true
          items:
            type: string
            enum: [scope-change, new-evidence, dependency-change, capacity-change, quality-finding, incident, estimate-correction, hold-change, other]
        impact: {type: string, minLength: 1, maxLength: 4000, pattern: '\S'}
    TargetInput:
      type: object
      additionalProperties: false
      required: [target_at, reason]
      properties:
        target_at: {type: string, format: date-time}
        reason: {type: string, minLength: 1, maxLength: 4000, pattern: '\S'}
    DeadlineInput:
      type: object
      additionalProperties: false
      required: [deadline_at, source, description]
      properties:
        deadline_at: {type: string, format: date-time}
        source: {type: string, enum: [contract, launch-window, demonstration, regulation, other]}
        description: {type: string, minLength: 1, maxLength: 4000, pattern: '\S'}
...
    Forecast:
      type: object
      additionalProperties: false
      required: [id, organization_id, project_id, scope, p50_at, p90_at, review_after, basis, assumptions, reason_codes, impact, created_by, created_at, current, stale, attention_only]
      properties:
        id: {type: string, format: uuid}
        organization_id: {type: string, format: uuid}
        project_id: {type: string, format: uuid}
        deliverable_id: {type: [string, 'null'], format: uuid}
        scope: {type: string, enum: [project, deliverable]}
        p50_at: {type: string, format: date-time}
        p90_at: {type: string, format: date-time}
        review_after: {type: string, format: date-time}
        basis: {type: string}
        assumptions: {type: array, items: {type: string}}
        reason_codes: {type: array, items: {type: string}}
        impact: {type: string}
        supersedes_id: {type: [string, 'null'], format: uuid}
        created_by: {type: string, format: uuid}
        created_at: {type: string, format: date-time}
        current: {type: boolean}
        stale: {type: boolean}
        attention_only: {type: boolean, const: true}
    Target:
      type: object
      additionalProperties: false
      required: [id, organization_id, project_id, target_at, reason, created_by, created_at, current, missed, attention_only]
      properties:
        id: {type: string, format: uuid}
        organization_id: {type: string, format: uuid}
        project_id: {type: string, format: uuid}
        target_at: {type: string, format: date-time}
        reason: {type: string}
        supersedes_id: {type: [string, 'null'], format: uuid}
        created_by: {type: string, format: uuid}
        created_at: {type: string, format: date-time}
        current: {type: boolean}
        missed: {type: boolean}
        attention_only: {type: boolean, const: true}
    Deadline:
      type: object
      additionalProperties: false
      required: [id, organization_id, project_id, deadline_at, source, description, created_by, created_at, current, passed, attention_only]
      properties:
        id: {type: string, format: uuid}
        organization_id: {type: string, format: uuid}
        project_id: {type: string, format: uuid}
        deadline_at: {type: string, format: date-time}
        source: {type: string, enum: [contract, launch-window, demonstration, regulation, other]}
        description: {type: string}
        supersedes_id: {type: [string, 'null'], format: uuid}
        created_by: {type: string, format: uuid}
        created_at: {type: string, format: date-time}
        current: {type: boolean}
        passed: {type: boolean}
        attention_only: {type: boolean, const: true}
    PromotionResult:
      type: object
      additionalProperties: false
      required: [project, decision, deliverables]
      properties:
        project: {$ref: '#/components/schemas/Project'}
        decision: {$ref: '#/components/schemas/Decision'}
        deliverables: {type: array, items: {$ref: '#/components/schemas/Deliverable'}}
...