# projects/workplane/internal/api/generated/server.gen.go
// Code generated by scripts/generate.py from contracts/openapi.yaml; DO NOT EDIT.
// Contract SHA-256: 9674850c6335a55d5ec4b11534710ac07784aa76b71eed21ca8b667e0518809b
package generated

import (
	"context"
	"encoding/json"
	"io"
	"net/http"
	"strings"
)

type OperationID string

const (
	OperationActivateProject          OperationID = "activateProject"
	OperationCreateDeliverable        OperationID = "createDeliverable"
	OperationCreateProject            OperationID = "createProject"
	OperationGetDeliverable           OperationID = "getDeliverable"
	OperationGetProject               OperationID = "getProject"
	OperationHoldProject              OperationID = "holdProject"
	OperationListDeliverableForecasts OperationID = "listDeliverableForecasts"
	OperationListDeliverables         OperationID = "listDeliverables"
	OperationListProjectActivity      OperationID = "listProjectActivity"
	OperationListProjectForecasts     OperationID = "listProjectForecasts"
	OperationLogin                    OperationID = "login"
	OperationPromoteProject           OperationID = "promoteProject"
	OperationRecordDecision           OperationID = "recordDecision"
	OperationReforecastDeliverable    OperationID = "reforecastDeliverable"
	OperationReforecastProject        OperationID = "reforecastProject"
	OperationResumeProject            OperationID = "resumeProject"
	OperationReviseDeliverable        OperationID = "reviseDeliverable"
	OperationSetProjectDeadline       OperationID = "setProjectDeadline"
	OperationSetProjectTarget         OperationID = "setProjectTarget"

	RequestBodyMaxBytes                                   = 65536
	IdempotencyKeyMinLength                               = 16
	IdempotencyKeyMaxLength                               = 128
	LoginRequestEmailMinLength                            = 1
	LoginRequestEmailMaxLength                            = 320
	LoginRequestPasswordMinLength                         = 1
	LoginRequestPasswordMaxLength                         = 1024
	CreateExplorationProjectTitleMinLength                = 1
	CreateExplorationProjectTitleMaxLength                = 200
	CreateExplorationProjectOutcomeMinLength              = 1
	CreateExplorationProjectOutcomeMaxLength              = 2000
	CreateExplorationProjectHypothesisMinLength           = 1
	CreateExplorationProjectHypothesisMaxLength           = 2000
	CreateExplorationProjectFalsifierMinLength            = 1
	CreateExplorationProjectFalsifierMaxLength            = 2000
	CreateExplorationProjectDecisionCriteriaMinItems      = 1
	CreateExplorationProjectDecisionCriteriaMaxItems      = 32
	CreateExplorationProjectDecisionCriteriaItemMinLength = 1
	CreateExplorationProjectDecisionCriteriaItemMaxLength = 500
	CreateExplorationProjectExperimentBoundMinLength      = 1
	CreateExplorationProjectExperimentBoundMaxLength      = 1000
	RecordDecisionQuestionMinLength                       = 1
	RecordDecisionQuestionMaxLength                       = 2000
	RecordDecisionChoiceMinLength                         = 1
	RecordDecisionChoiceMaxLength                         = 2000
	RecordDecisionAlternativesMaxItems                    = 32
	RecordDecisionAlternativesItemMaxLength               = 1000
	RecordDecisionRationaleMinLength                      = 1
	RecordDecisionRationaleMaxLength                      = 4000
	RecordDecisionEvidenceMaxItems                        = 64
	RecordDecisionEvidenceItemMaxLength                   = 2000
	RecordDecisionConsequencesMaxItems                    = 32
	RecordDecisionConsequencesItemMaxLength               = 2000
	LifecycleReasonRequestReasonMinLength                 = 1
	LifecycleReasonRequestReasonMaxLength                 = 4000
	DeliverableInputTitleMinLength                        = 1
	DeliverableInputTitleMaxLength                        = 200
	DeliverableInputDescriptionMinLength                  = 1
	DeliverableInputDescriptionMaxLength                  = 4000
	DeliverableInputAcceptanceCriteriaMinItems            = 1
	DeliverableInputAcceptanceCriteriaMaxItems            = 32
	DeliverableInputAcceptanceCriteriaItemMinLength       = 1
	DeliverableInputAcceptanceCriteriaItemMaxLength       = 1000
	PromoteProjectRequestResidualUncertaintyMinLength     = 1
	PromoteProjectRequestResidualUncertaintyMaxLength     = 4000
	PromoteProjectRequestPriorityRationaleMinLength       = 1
	PromoteProjectRequestPriorityRationaleMaxLength       = 4000
	PromoteProjectRequestDeliverablesMinItems             = 1
	PromoteProjectRequestDeliverablesMaxItems             = 32
	ForecastInputBasisMinLength                           = 1
	ForecastInputBasisMaxLength                           = 4000
	ForecastInputAssumptionsMinItems                      = 1
	ForecastInputAssumptionsMaxItems                      = 32
	ForecastInputAssumptionsItemMinLength                 = 1
	ForecastInputAssumptionsItemMaxLength                 = 1000
	ForecastInputReasonCodesMinItems                      = 1
	ForecastInputReasonCodesMaxItems                      = 9
	ForecastInputImpactMinLength                          = 1
	ForecastInputImpactMaxLength                          = 4000
	TargetInputReasonMinLength                            = 1
	TargetInputReasonMaxLength                            = 4000
	DeadlineInputDescriptionMinLength                     = 1
	DeadlineInputDescriptionMaxLength                     = 4000
)

type RequestSecurity struct {
	SessionCookie string
	CSRFToken     string
	BearerToken   string
}

type VersionETag string

type LoginRequest struct {
	Email    string `json:"email"`
	Password string `json:"password"`
}

type CreateExplorationProject struct {
	Title            string   `json:"title"`
	Outcome          string   `json:"outcome"`
	Hypothesis       string   `json:"hypothesis"`
	Falsifier        string   `json:"falsifier"`
	DecisionCriteria []string `json:"decision_criteria"`
	ExperimentBound  string   `json:"experiment_bound"`
}

type RecordDecision struct {
	Kind         string   `json:"kind"`
	Question     string   `json:"question"`
	Choice       string   `json:"choice"`
	Alternatives []string `json:"alternatives"`
	Rationale    string   `json:"rationale"`
	Evidence     []string `json:"evidence"`
	Consequences []string `json:"consequences"`
}

type LifecycleReasonRequest struct {
	Reason string `json:"reason"`
}

type DeliverableInput struct {
	Title              string   `json:"title"`
	Description        string   `json:"description"`
	Required           bool     `json:"required"`
	Weight             int64    `json:"weight"`
	State              string   `json:"state"`
	AcceptanceCriteria []string `json:"acceptance_criteria"`
}

type PromoteProjectRequest struct {
	Decision            PromotionDecision  `json:"decision"`
	ResidualUncertainty string             `json:"residual_uncertainty"`
	PriorityRationale   string             `json:"priority_rationale"`
	Deliverables        []DeliverableInput `json:"deliverables"`
}

type ForecastInput struct {
	P50At       string   `json:"p50_at"`
	P90At       string   `json:"p90_at"`
	ReviewAfter string   `json:"review_after"`
	Basis       string   `json:"basis"`
	Assumptions []string `json:"assumptions"`
	ReasonCodes []string `json:"reason_codes"`
	Impact      string   `json:"impact"`
}

type TargetInput struct {
	TargetAt string `json:"target_at"`
	Reason   string `json:"reason"`
}

type DeadlineInput struct {
	DeadlineAt  string `json:"deadline_at"`
	Source      string `json:"source"`
	Description string `json:"description"`
}

type PromotionDecision struct {
	Question     string   `json:"question"`
	Choice       string   `json:"choice"`
	Alternatives []string `json:"alternatives"`
	Rationale    string   `json:"rationale"`
	Evidence     []string `json:"evidence"`
	Consequences []string `json:"consequences"`
}

type Session struct {
	ActorID   string `json:"actor_id"`
	ActorKind string `json:"actor_kind"`
	ExpiresAt string `json:"expires_at"`
	CSRFToken string `json:"csrf_token"`
}

type Request struct {
	HTTPRequest     *http.Request
	Body            json.RawMessage
	IdempotencyKey  string
	ExpectedVersion VersionETag
	Security        RequestSecurity
}

type ResponseHeaders struct {
	ETag       VersionETag
	SetCookie  string
	XRequestID string
}

type Response struct {
	Status      int
	ContentType string
	Headers     ResponseHeaders
	Body        any
}

type Handler interface {
	ActivateProject(context.Context, Request) (Response, error)
	CreateDeliverable(context.Context, Request) (Response, error)
	CreateProject(context.Context, Request) (Response, error)
	GetDeliverable(context.Context, Request) (Response, error)
	GetProject(context.Context, Request) (Response, error)
	HoldProject(context.Context, Request) (Response, error)
	ListDeliverableForecasts(context.Context, Request) (Response, error)
	ListDeliverables(context.Context, Request) (Response, error)
	ListProjectActivity(context.Context, Request) (Response, error)
	ListProjectForecasts(context.Context, Request) (Response, error)
	Login(context.Context, Request) (Response, error)
	PromoteProject(context.Context, Request) (Response, error)
	RecordDecision(context.Context, Request) (Response, error)
	ReforecastDeliverable(context.Context, Request) (Response, error)
	ReforecastProject(context.Context, Request) (Response, error)
	ResumeProject(context.Context, Request) (Response, error)
	ReviseDeliverable(context.Context, Request) (Response, error)
	SetProjectDeadline(context.Context, Request) (Response, error)
	SetProjectTarget(context.Context, Request) (Response, error)
}

func RegisterHandlers(mux *http.ServeMux, handler Handler) {
	mux.HandleFunc("POST /api/v1/projects/{project_id}/activate", adapt(handler.ActivateProject))
	mux.HandleFunc("POST /api/v1/projects/{project_id}/deliverables", adapt(handler.CreateDeliverable))
	mux.HandleFunc("POST /api/v1/orgs/{org_id}/projects", adapt(handler.CreateProject))
	mux.HandleFunc("GET /api/v1/deliverables/{id}", adapt(handler.GetDeliverable))
	mux.HandleFunc("GET /api/v1/projects/{project_id}", adapt(handler.GetProject))
	mux.HandleFunc("POST /api/v1/projects/{project_id}/hold", adapt(handler.HoldProject))
	mux.HandleFunc("GET /api/v1/deliverables/{deliverable_id}/forecasts", adapt(handler.ListDeliverableForecasts))
	mux.HandleFunc("GET /api/v1/projects/{project_id}/deliverables", adapt(handler.ListDeliverables))
	mux.HandleFunc("GET /api/v1/projects/{project_id}/activity", adapt(handler.ListProjectActivity))
	mux.HandleFunc("GET /api/v1/projects/{project_id}/forecasts", adapt(handler.ListProjectForecasts))
	mux.HandleFunc("POST /api/v1/session/login", adapt(handler.Login))
	mux.HandleFunc("POST /api/v1/projects/{project_id}/promote", adapt(handler.PromoteProject))
	mux.HandleFunc("POST /api/v1/projects/{project_id}/decisions", adapt(handler.RecordDecision))
	mux.HandleFunc("POST /api/v1/deliverables/{deliverable_id}/forecasts", adapt(handler.ReforecastDeliverable))
	mux.HandleFunc("POST /api/v1/projects/{project_id}/forecasts", adapt(handler.ReforecastProject))
	mux.HandleFunc("POST /api/v1/projects/{project_id}/resume", adapt(handler.ResumeProject))
	mux.HandleFunc("PATCH /api/v1/deliverables/{id}", adapt(handler.ReviseDeliverable))
	mux.HandleFunc("POST /api/v1/projects/{project_id}/deadline", adapt(handler.SetProjectDeadline))
	mux.HandleFunc("POST /api/v1/projects/{project_id}/target", adapt(handler.SetProjectTarget))
}

type operationHandler func(context.Context, Request) (Response, error)

func adapt(handler operationHandler) http.HandlerFunc {
	return func(writer http.ResponseWriter, request *http.Request) {
		var body json.RawMessage
		if request.Body != nil {
			defer request.Body.Close()
			value, err := io.ReadAll(io.LimitReader(request.Body, RequestBodyMaxBytes+1))
			if err != nil {
				writeAdapterProblem(writer, http.StatusBadRequest, "invalid_request")
				return
			}
			if len(value) > RequestBodyMaxBytes {
				writeAdapterProblem(writer, http.StatusRequestEntityTooLarge, "invalid_request")
				return
			}
			body = value
		}
		sessionCookie := ""
		if cookie, err := request.Cookie("workplane_session"); err == nil {
			sessionCookie = cookie.Value
		}
		bearerToken := ""
		if authorization := request.Header.Get("Authorization"); strings.HasPrefix(authorization, "Bearer ") {
			bearerToken = strings.TrimPrefix(authorization, "Bearer ")
		}
		response, err := handler(request.Context(), Request{
			HTTPRequest:     request,
			Body:            body,
			IdempotencyKey:  request.Header.Get("Idempotency-Key"),
			ExpectedVersion: VersionETag(request.Header.Get("If-Match")),
			Security: RequestSecurity{
				SessionCookie: sessionCookie,
				CSRFToken:     request.Header.Get("X-CSRF-Token"),
				BearerToken:   bearerToken,
			},
		})
		if err != nil {
			writeAdapterProblem(writer, http.StatusInternalServerError, "service_unavailable")
			return
		}
		contentType := response.ContentType
		if contentType == "" {
			contentType = "application/json"
		}
		writer.Header().Set("Content-Type", contentType)
		writer.Header().Set("Cache-Control", "no-store")
		writer.Header().Set("X-Content-Type-Options", "nosniff")
		writer.Header().Set("Referrer-Policy", "no-referrer")
		writer.Header().Set("Content-Security-Policy", "default-src 'self'; object-src 'none'; frame-ancestors 'none'; base-uri 'none'")
		if response.Headers.ETag != "" {
			writer.Header().Set("ETag", string(response.Headers.ETag))
		}
		if response.Headers.SetCookie != "" {
			writer.Header().Set("Set-Cookie", response.Headers.SetCookie)
		}
		if response.Headers.XRequestID != "" {
			writer.Header().Set("X-Request-ID", response.Headers.XRequestID)
		}
		writer.WriteHeader(response.Status)
		_ = json.NewEncoder(writer).Encode(response.Body)
	}
}

func writeAdapterProblem(writer http.ResponseWriter, status int, code string) {
	writer.Header().Set("Content-Type", "application/problem+json")
	writer.Header().Set("Cache-Control", "no-store")
	writer.WriteHeader(status)
	_ = json.NewEncoder(writer).Encode(map[string]any{
		"type":       "https://workplane.local/problems/" + code,
		"title":      http.StatusText(status),
		"status":     status,
		"code":       code,
		"request_id": "adapter",
	})
}